requirements.txt
Flask>=2.3.0
opencv-python>=4.8.0
tensorflow>=2.16.0
numpy>=1.24.0
Pillow>=10.0.0